requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.32.1
langchain-openai==0.2.9
langgraph==0.6.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3
httpx==0.28.1
jinja2==3.1.5
python-multipart==0.0.19
aiofiles==24.1.0
redis==5.2.0
numpy==2.2.0
sentence-transformers==3.3.1
pymongo==4.10.1
docling==2.47.0
python-docx==1.1.2
openpyxl==3.1.5
PyPDF2==3.0.1
tiktoken==0.8.0
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0